+++ libs/gesha-core/src/conversions/mod.rs
@@ -2,18 +2,4 @@ mod error;
 pub use error::{Error, Result};
 
 mod converter;
-pub use converter::{Converter, TestDefinition};
-
-mod test_case;
-pub use test_case::TestCase;
-
-mod test_case_map;
-pub use test_case_map::TestCaseMap;
-
-mod test_suite;
-pub use test_suite::TestSuite;
-
-mod test_runner;
-pub use test_runner::TestRunner;
-
-pub mod v3_0;
+pub use converter::Converter;
