+++ libs/gesha-core/src/lib.rs
@@ -4,3 +4,5 @@ pub use crate::error::{DiffStyle, Error, ErrorTheme, Result};
 pub mod conversions;
 mod io;
 pub mod trace;
+
+pub mod testing;
