+++ libs/gesha-core/src/testing/test_definition.rs
@@ -1,0 +1,30 @@
+use crate::conversions::Converter;
+use crate::testing::{TestCase, TestSuite};
+use crate::Error::UnknownTestCase;
+
+pub trait TestDefinition: Converter {
+    /// List all test suites for this definition.
+    fn test_suites(&self) -> Vec<TestSuite<Self>>;
+
+    /// Get the code of the test suite.
+    fn test_suite_code(&self, suite: &TestSuite<Self>) -> Self::TargetType;
+
+    /// List all test cases for this definition.
+    fn list_test_cases(&self) -> Vec<TestCase<Self>> {
+        self.test_suites()
+            .into_iter()
+            .flat_map(|suite| suite.test_cases)
+            .collect()
+    }
+
+    /// Get a test case by its path.
+    /// If the test case does not exist, return an error.
+    fn require_test_case(&self, path: &str) -> Result<TestCase<Self>, crate::Error> {
+        self.list_test_cases()
+            .into_iter()
+            .find(|case| case.schema.as_os_str() == path)
+            .ok_or_else(|| UnknownTestCase {
+                path: path.to_string(),
+            })
+    }
+}
