+++ libs/gesha-core/src/testing/test_suite.rs
@@ -1,4 +1,4 @@
-use crate::conversions::TestCase;
+use crate::testing::TestCase;
 use std::path::PathBuf;
 
 #[derive(Debug)]
